moved FAQPanel to components, storybooked it
