+++ client/src/components/FAQPanel/FAQPanel.yaml
@@ -1,0 +1,20 @@
+question1_q:
+  en: |
+    Question 1?
+question1_a:
+  en: |
+    Answer 1.
+
+question2_q:
+  en: |
+    Question 2?
+question2_a:
+  en: |
+    Answer 2.
+
+question3_q:
+  en: |
+    Question 3?
+question3_a:
+  en: |
+    Answer 3.
