+++ client/src/panels/panel_declarations/misc/key_concept_panels/key_concept_panels.js
@@ -1,32 +1,9 @@
 import _ from "lodash";
 import React from "react";
 
-import { SomeThingsToKeepInMind } from "src/panels/panel_declarations/common_panel_components";
 import { declare_panel } from "src/panels/panel_declarations/common_panel_utils";
 
-import {
-  create_text_maker_component,
-  KeyConceptList,
-} from "src/components/index";
-
-import { Gov } from "src/models/organizational_entities";
-
-import est_lang from "src/EstimatesComparison/estimates_comparison_questions.yaml";
-
-import common_lang from "./common_questions.yaml";
-import fin_lang from "./financial_questions.yaml";
-import ppl_lang from "./people_questions.yaml";
-import results_lang from "./results_questions.yaml";
-import tag_lang from "./tagging_questions.yaml";
-
-const { TM } = create_text_maker_component([
-  common_lang,
-  fin_lang,
-  ppl_lang,
-  results_lang,
-  tag_lang,
-  est_lang,
-]);
+import { FAQPanel } from "src/components/index";
 
 const common_panel_config = {
   is_static: true,
@@ -34,33 +11,6 @@ const common_panel_config = {
   source: false,
   calculate: _.constant(true),
 };
-
-export class FAQPanel extends React.Component {
-  render() {
-    const {
-      rendered_q_a_keys,
-      is_initially_expanded,
-      background_color,
-      subject,
-    } = this.props;
-
-    let subj = _.isUndefined(subject) ? Gov : subject;
-
-    return (
-      <SomeThingsToKeepInMind
-        is_initially_expanded={is_initially_expanded}
-        background_color={background_color}
-      >
-        <KeyConceptList
-          question_answer_pairs={_.map(rendered_q_a_keys, (base_text_key) => [
-            <TM key={"q"} k={base_text_key + "_q"} args={{ subj }} />,
-            <TM key={"a"} k={base_text_key + "_a"} args={{ subj }} />,
-          ])}
-        />
-      </SomeThingsToKeepInMind>
-    );
-  }
-}
 
 const curried_render = ({ q_a_keys }) =>
   function ({ calculations: { subject } }) {
